Çeşitli dosyalarda iyileştirmeler ve düzeltmeler yapıldı: ACPI güç kapatma işlevi güncellendi, bellek yönetimi ve PCI yapılandırması için yeni işlevler eklendi, hata ayıklama logları geliştirildi, ekran başlangıç logları eklendi, ve derleyici bayrakları optimize edildi.

+++ kernel/memory/pmm.c
@@ -527,7 +527,7 @@ void *pmm_alloc(size_t sizeInKB)
         return (void *)(uintptr_t)alloc_region->base;
     }
 
-    ERROR("pmm_alloc: no suitable block found for sizeInKB=%zu", sizeInKB);
+    ERROR("pmm_alloc: no suitable block found for sizeInKB=%08u", (uint32_t)sizeInKB);
     return NULL; // uygun blok yok
 }
